+++ sdk/core/azure-core/src/test/java/com/azure/core/util/CoreUtilsTests.java
@@ -690,4 +690,9 @@ public class CoreUtilsTests {
         assertThrows(IllegalArgumentException.class,
             () -> CoreUtils.addShutdownHookSafely(Executors.newSingleThreadExecutor(), Duration.ZERO));
     }
+
+    @Test
+    public void addRuntimeShutdownHookSafelyWithNullThreadDoesNothing() {
+        assertNull(CoreUtils.addRuntimeShutdownHookSafely(null));
+    }
 }