+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added `CoreUtils.addRuntimeShutdownHookSafely` which is a more generic version of `CoreUtils.addShutdownHookSafely`. ([#38730](https://github.com/Azure/azure-sdk-for-java/pull/38730))
+
 ### Breaking Changes
 
 ### Bugs Fixed